+++ microcontroller/src-atmel/borg/borg-apiSimulator/tetris/view.c
@@ -166,20 +166,20 @@ void tetris_view_update(tetris_view_t *pV)
 void tetris_view_showResults(tetris_view_t *pV)
 {
 	char pszResults[64];
-	uint32_t nScore = tetris_logic_getScore(pV->pLogic);
-	uint32_t nHighscore = tetris_logic_getHighscore(pV->pLogic);
+	uint16_t nScore = tetris_logic_getScore(pV->pLogic);
+	uint16_t nHighscore = tetris_logic_getHighscore(pV->pLogic);
 	uint8_t nLines = tetris_logic_getLines(pV->pLogic);
 	
 	if (nScore <= nHighscore)
 	{
 		snprintf(pszResults, 64 * sizeof(char),
-			"</#Lines %d    Your Score %d    Highscore %d",
+			"</#Lines %u    Your Score %u    Highscore %u",
 			nLines, nScore, nHighscore);
 	}
 	else
 	{
 		snprintf(pszResults, 64 * sizeof(char),
-			"</#Lines %d    New Highscore %d", nLines, nScore);
+			"</#Lines %u    New Highscore %u", nLines, nScore);
 	}
 	
 	scrolltext(pszResults);
